+++ packages/preact/src/methods/insert.ts
@@ -18,6 +18,8 @@ import {
   setFieldState,
   sortArrayPathIndex,
   validateIfRequired,
+  getFieldNames,
+  getFieldArrayNames,
 } from '../utils';
 
 /**
@@ -79,8 +81,7 @@ export function insert<
             ) as T;
 
           // Move fields that come after new item one index further
-          form.internal.fieldNames
-            .peek()
+          getFieldNames(form)
             .filter(filterName)
             .sort(sortArrayPathIndex(name))
             .reverse()
@@ -93,8 +94,7 @@ export function insert<
             });
 
           // Move field arrays that come after new item one index further
-          form.internal.fieldArrayNames
-            .peek()
+          getFieldArrayNames(form)
             .filter(filterName)
             .sort(sortArrayPathIndex(name))
             .reverse()