+++ packages/preact/src/methods/remove.ts
@@ -6,8 +6,10 @@ import type {
   ResponseData,
 } from '../types';
 import {
+  getFieldArrayNames,
   getFieldArrayState,
   getFieldArrayStore,
+  getFieldNames,
   getFieldState,
   getPathIndex,
   setFieldArrayState,
@@ -65,8 +67,7 @@ export function remove<
 
       batch(() => {
         // Move state of each field after the removed index back by one index
-        form.internal.fieldNames
-          .peek()
+        getFieldNames(form)
           .filter(filterName)
           .sort(sortArrayPathIndex(name))
           .forEach((fieldName) => {
@@ -78,8 +79,7 @@ export function remove<
           });
 
         // Move state of each field array after the removed index back by one index
-        form.internal.fieldArrayNames
-          .peek()
+        getFieldArrayNames(form)
           .filter(filterName)
           .sort(sortArrayPathIndex(name))
           .forEach((fieldArrayName) => {