+++ packages/react/src/utils/getFilteredNames.ts
@@ -6,6 +6,8 @@ import type {
   Maybe,
   ResponseData,
 } from '../types';
+import { getFieldArrayNames } from './getFieldArrayNames';
+import { getFieldNames } from './getFieldNames';
 
 /**
  * Returns a tuple with filtered field and field array names. For each
@@ -15,6 +17,7 @@ import type {
  *
  * @param form The form of the fields.
  * @param arg2 The name of the fields.
+ * @param shouldValid Whether to be valid.
  *
  * @returns A tuple with filtered names.
  */
@@ -29,11 +32,12 @@ export function getFilteredNames<
     | FieldArrayPath<TFieldValues>
     | (FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>)[]
     | TOptions
-  >
+  >,
+  shouldValid?: Maybe<boolean>
 ): [FieldPath<TFieldValues>[], FieldArrayPath<TFieldValues>[]] {
   // Get all field and field array names of form
-  const allFieldNames = form.internal.fieldNames.peek();
-  const allFieldArrayNames = form.internal.fieldArrayNames.peek();
+  const allFieldNames = getFieldNames(form, shouldValid);
+  const allFieldArrayNames = getFieldArrayNames(form, shouldValid);
 
   // If names are specified, filter and return them
   if (typeof arg2 === 'string' || Array.isArray(arg2)) {