+++ packages/qwik/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the library will be documented in this file.
 - Require validation when using `formAction$` (issue #57)
 - Refactor `getFilteredNames` util
 - Refactor path index sorting in `insert`, `move` and `remove` method
+- Fix bug by removing invalid field array items (issue #85)
 
 ## v0.17.1 (June 17, 2023)