+++ packages/solid/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the library will be documented in this file.
 - Rename `noValidate` in `Form` component to `novalidate` (issue #92)
 - Refactor `getFilteredNames` util
 - Refactor path index sorting in `insert`, `move` and `remove` method
+- Fix bug by removing invalid field array items (issue #85)
 
 ## v0.17.2 (June 17, 2023)